Arm Code
This is most likely not going to work, but try it out. There are some things that you can try, and refer back to my relic recovery code. I feel as though there was a reason as to why I set my target to negative on the arm, so try that out, as that's probably going to be your quickest fix